support mow

+++ regression-test/suites/insert_p0/transaction/sub_txn_mow.groovy
@@ -15,11 +15,15 @@
 // specific language governing permissions and limitations
 // under the License.
 
-suite("sub_txn_mor") {
+suite("sub_txn_mow") {
+    if (isCloudMode()) {
+        logger.info("txn load for mow tables is not supported in cloud mode")
+        return
+    }
     sql """ set enable_query_in_transaction_load = true """
     // case 1
-    def table_txn = "sub_txn_mor"
-    def table_normal = "sub_txn_mor_n"
+    def table_txn = "sub_txn_mow"
+    def table_normal = "sub_txn_mow_n"
     for (def i in 1..3) {
         for (def prefix: [table_normal, table_txn]) {
             sql """ drop table if exists ${prefix}_${i} """
@@ -32,8 +36,8 @@ suite("sub_txn_mor") {
             UNIQUE KEY(`id`)
             DISTRIBUTED BY HASH(`id`) BUCKETS 1
             PROPERTIES (
-                "enable_unique_key_merge_on_write" = "false",
-                "replication_num" = "1"
+                "replication_num" = "1",
+                "enable_mow_light_delete" = "true"
             );
             """
             if (i == 1) {
@@ -47,28 +51,31 @@ suite("sub_txn_mor") {
         }
 
         sql """ insert into ${prefix}_3 select * from ${prefix}_2; """
-        order_qt_select_1 """ select * from ${prefix}_3; """
+        // order_qt_select_1 """ select * from ${prefix}_3; """
 
         sql """ insert into ${prefix}_3 select * from ${prefix}_1; """
-        order_qt_select_2 """ select * from ${prefix}_3; """
+        // order_qt_select_2 """ select * from ${prefix}_3; """
 
         sql """ insert into ${prefix}_2 select * from ${prefix}_3; """
-        order_qt_select_3 """ select * from ${prefix}_2; """
+        // order_qt_select_3 """ select * from ${prefix}_2; """
 
         sql """ insert into ${prefix}_1 select * from ${prefix}_2; """
-        order_qt_select_4 """ select * from ${prefix}_1; """
+        // order_qt_select_4 """ select * from ${prefix}_1; """
+
+        sql """ insert into ${prefix}_2 select * from ${prefix}_1; """
+        // order_qt_select_5 """ select * from ${prefix}_2; """
 
         if (prefix == table_txn) {
             sql """ commit; """
         }
-        order_qt_select_51 """ select * from ${prefix}_1; """
-        order_qt_select_52 """ select * from ${prefix}_2; """
-        order_qt_select_53 """ select * from ${prefix}_3; """
+        order_qt_select_61 """ select * from ${prefix}_1; """
+        order_qt_select_62 """ select * from ${prefix}_2; """
+        order_qt_select_63 """ select * from ${prefix}_3; """
     }
 
     // case 2: insert with partition
-    table_txn = "sub_txn_mor_p"
-    table_normal = "sub_txn_mor_pn"
+    table_txn = "sub_txn_mow_p"
+    table_normal = "sub_txn_mow_pn"
     for (def i in 1..3) {
         for (def prefix: [table_normal, table_txn]) {
             sql """ drop table if exists ${prefix}_${i} """
@@ -101,22 +108,22 @@ suite("sub_txn_mor") {
         }
 
         sql """ insert into ${prefix}_3 PARTITION(p_1_11) select * from ${prefix}_2; """
-        order_qt_par_1 """ select * from ${prefix}_3; """
+        // order_qt_par_1 """ select * from ${prefix}_3; """
 
         sql """ insert into ${prefix}_2 PARTITION(p_1_11) select * from ${prefix}_3; """
-        order_qt_par_2 """ select * from ${prefix}_2; """
+        // order_qt_par_2 """ select * from ${prefix}_2; """
 
         sql """ insert into ${prefix}_3 select * from ${prefix}_1; """
-        order_qt_par_3 """ select * from ${prefix}_3; """
+        // order_qt_par_3 """ select * from ${prefix}_3; """
 
         sql """ insert into ${prefix}_2 select * from ${prefix}_3; """
-        order_qt_par_4 """ select * from ${prefix}_2; """
+        // order_qt_par_4 """ select * from ${prefix}_2; """
 
         sql """ insert into ${prefix}_3 PARTITION(p_1_11) select * from ${prefix}_2; """
-        order_qt_par_5 """ select * from ${prefix}_3; """
+        // order_qt_par_5 """ select * from ${prefix}_3; """
 
         sql """ insert into ${prefix}_2 PARTITION(p_11_21) select * from ${prefix}_3; """
-        order_qt_par_6 """ select * from ${prefix}_2; """
+        // order_qt_par_6 """ select * from ${prefix}_2; """
 
         if (prefix == table_txn) {
             sql """ commit; """
@@ -128,30 +135,30 @@ suite("sub_txn_mor") {
     sql """ set enable_insert_strict = true """
 
     // case 3: delete command
-    table_txn = "sub_txn_mor"
-    table_normal = "sub_txn_mor_n"
+    table_txn = "sub_txn_mow"
+    table_normal = "sub_txn_mow_n"
     for (def prefix: [table_normal, table_txn]) {
         if (prefix == table_txn) {
             sql """ begin; """
         }
 
         sql """ delete from ${prefix}_3 where id > 1; """
-        order_qt_del_1 """ select * from ${prefix}_3; """
+        // order_qt_del_1 """ select * from ${prefix}_3; """
 
         sql """ insert into ${prefix}_2 select * from ${prefix}_3; """
-        order_qt_del_2 """ select * from ${prefix}_2; """
+        // order_qt_del_2 """ select * from ${prefix}_2; """
 
         sql """ insert into ${prefix}_3 select * from ${prefix}_2; """
-        order_qt_del_3 """ select * from ${prefix}_3; """
+        // order_qt_del_3 """ select * from ${prefix}_3; """
 
         sql """ update ${prefix}_3 set score = score + 100 where id = 1; """
-        order_qt_del_4 """ select * from ${prefix}_3; """
+        // order_qt_del_4 """ select * from ${prefix}_3; """
 
         sql """ delete from ${prefix}_3 where id < 2; """
-        order_qt_del_5 """ select * from ${prefix}_3; """
+        // order_qt_del_5 """ select * from ${prefix}_3; """
 
         sql """ insert into ${prefix}_2 select * from ${prefix}_3; """
-        order_qt_del_6 """ select * from ${prefix}_2; """
+        // order_qt_del_6 """ select * from ${prefix}_2; """
 
         if (prefix == table_txn) {
             sql """ commit; """
